refactor(remote): keep attach watchdog policy private
The watchdog only serves the attach process driver, so keeping its state
machine private avoids widening the remote library surface.
