+++ crates/rimz/src/remote/tests.rs
@@ -993,45 +993,3 @@ fn attached_exit_uses_control_liveness_instead_of_mux_status() {
         }
     );
 }
-
-#[test]
-fn remote_room_watchdog_arms_on_work_and_fires_on_empty() {
-    let mut watchdog = RemoteRoomWatchdog::default();
-
-    assert_eq!(
-        watchdog.advance(Duration::ZERO, None),
-        RemoteRoomWatchdogAction::Observe
-    );
-    assert_eq!(
-        watchdog.advance(Duration::ZERO, Some(RemoteRoomObservation::WorkPanePresent),),
-        RemoteRoomWatchdogAction::Wait(REMOTE_ROOM_WATCHDOG_INTERVAL)
-    );
-    assert_eq!(
-        watchdog.advance(REMOTE_ROOM_WATCHDOG_INTERVAL, None),
-        RemoteRoomWatchdogAction::Observe
-    );
-    assert_eq!(
-        watchdog.advance(
-            REMOTE_ROOM_WATCHDOG_INTERVAL,
-            Some(RemoteRoomObservation::NoWorkPanes),
-        ),
-        RemoteRoomWatchdogAction::RoomEnded
-    );
-}
-
-#[test]
-fn remote_room_watchdog_fails_open_until_work_is_observed() {
-    let mut watchdog = RemoteRoomWatchdog::default();
-
-    assert_eq!(
-        watchdog.advance(Duration::ZERO, Some(RemoteRoomObservation::Unknown),),
-        RemoteRoomWatchdogAction::Wait(REMOTE_ROOM_WATCHDOG_INTERVAL)
-    );
-    assert_eq!(
-        watchdog.advance(
-            REMOTE_ROOM_WATCHDOG_INTERVAL,
-            Some(RemoteRoomObservation::NoWorkPanes),
-        ),
-        RemoteRoomWatchdogAction::Wait(REMOTE_ROOM_WATCHDOG_INTERVAL)
-    );
-}
